+++ src/MudBlazor.Docs/Pages/Components/SwipeArea/Examples/RealTimeSwipeExample.razor
@@ -1,6 +1,6 @@
 ﻿@namespace MudBlazor.Docs.Examples
 
-<MudSwipeArea @ref="_swipeArea" Class="mud-background border-dashed border-2 mud-border-primary" OnSwipeMove="HandleSwipeMove" OnSwipeEnd="Reset" OnSwipeLeave="@(() => { _swipeArea.Cancel(); Reset(); })" Style="width: 100%; height: 300px">
+<MudSwipeArea @ref="@_swipeArea" Class="mud-background border-dashed border-2 mud-border-primary" Style="width: 100%; height: 300px" OnSwipeMove="HandleSwipeMove" OnSwipeEnd="@Reset" OnSwipeLeave="@OnSwipeLeave">
     <MudText Class="pa-2" Style="user-select: none;" Typo="@Typo.body1">@($"{_swipeDirection[1].ToString()} {_swipeDirection[0].ToString()}")</MudText>
 </MudSwipeArea>
 
@@ -13,16 +13,16 @@
 <MudText Align="Align.Center">*Value resets if swipe goes out of area.</MudText>
 
 @code {
-    MudSwipeArea _swipeArea;
-    private List<SwipeDirection> _swipeDirection = new() { SwipeDirection.None, SwipeDirection.None };
     private double _leftSwipe;
     private double _rightSwipe;
     private double _topSwipe;
     private double _bottomSwipe;
+    private MudSwipeArea _swipeArea = null!;
+    private List<SwipeDirection> _swipeDirection = [SwipeDirection.None, SwipeDirection.None];
 
     private void HandleSwipeMove(MultiDimensionSwipeEventArgs e)
     {
-        _swipeDirection = new() { e.SwipeDirections[0], e.SwipeDirections[1] };
+        _swipeDirection = [e.SwipeDirections[0], e.SwipeDirections[1]];
         for (int i = 0; i < e.SwipeDirections.Count; i++)
         {
             if (e.SwipeDirections[i] == SwipeDirection.LeftToRight)
@@ -44,9 +44,15 @@
         }
     }
 
+    private void OnSwipeLeave()
+    {
+        _swipeArea.Cancel();
+        Reset();
+    }
+
     private void Reset()
     {
-        _swipeDirection = new() { SwipeDirection.None, SwipeDirection.None };
+        _swipeDirection = [SwipeDirection.None, SwipeDirection.None];
         _leftSwipe = 0;
         _rightSwipe = 0;
         _topSwipe = 0;
